requirements.txt
click==8.2.1
rich==14.0.0
google-cloud-storage>=2.0.0
google-auth>=2.0.0
google-cloud-compute>=1.0.0
pyyaml>=6.0.0
boto3>=1.26.0
pydantic>=2.0.0
 